Facility: move TextMessageParser and test to module stringparser, remove unused StubAction.
